+++ nielsen/example/web/webpack.config.js
@@ -2,15 +2,15 @@
 const path = require('path');
 const HTMLWebpackPlugin = require('html-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
+const NodePolyfillPlugin = require('node-polyfill-webpack-plugin');
 
-const projectDirectory = path.resolve(__dirname, '../..');
-const appDirectory = path.resolve(__dirname, '..');
+const projectDirectory = path.resolve(__dirname, '..');
 
 // A folder for any stub components we need in case there is no counterpart for it on react-native-web.
-const stubDirectory = path.resolve(appDirectory, './web/stub/');
+const stubDirectory = path.resolve(projectDirectory, './web/stub/');
 
 const HTMLWebpackPluginConfig = new HTMLWebpackPlugin({
-  template: path.resolve(appDirectory, './web/public/index.html'),
+  template: path.resolve(projectDirectory, './web/public/index.html'),
   filename: 'index.html',
   inject: 'body',
 });
@@ -31,9 +31,9 @@ const CopyWebpackPluginConfig = new CopyWebpackPlugin({
     },
     {
       // Copy CSS files
-      from: path.resolve(appDirectory, './web/public/*.css').replace(/\\/g, '/'),
+      from: path.resolve(projectDirectory, './web/public/*.css').replace(/\\/g, '/'),
       to: `[name][ext]`,
-    }
+    },
   ],
 });
 
@@ -68,15 +68,15 @@ const imageLoaderConfiguration = {
 module.exports = {
   entry: [
     // load any web API polyfills
-    // path.resolve(appDirectory, 'polyfills-web.js'),
+    // path.resolve(projectDirectory, 'polyfills-web.js'),
     // your web-specific entry file
-    path.resolve(appDirectory, 'index.web.tsx'),
+    path.resolve(projectDirectory, 'index.web.tsx'),
   ],
 
   // configures where the build ends up
   output: {
     filename: 'bundle.web.js',
-    path: path.resolve(appDirectory, outputLocation),
+    path: path.resolve(projectDirectory, outputLocation),
   },
 
   module: {
@@ -88,20 +88,22 @@ module.exports = {
       'react-native$': 'react-native-web',
       'react-native-url-polyfill': 'url-polyfill',
       'react-native-google-cast': path.resolve(stubDirectory, 'CastButtonStub'),
-      'react-native-web': path.resolve(appDirectory, 'node_modules/react-native-web'),
+      'react-native-web': path.resolve(projectDirectory, 'node_modules/react-native-web'),
+      'react-native-svg': 'react-native-svg-web',
 
       // Avoid duplicate react env.
-      react: path.resolve(appDirectory, 'node_modules/react'),
+      'react': path.resolve(projectDirectory, 'node_modules/react'),
+      'react-dom': path.resolve(projectDirectory, 'node_modules/react-dom')
     },
   },
-  plugins: [HTMLWebpackPluginConfig, CopyWebpackPluginConfig],
+  plugins: [HTMLWebpackPluginConfig, CopyWebpackPluginConfig, new NodePolyfillPlugin()],
   devServer: {
     // Tells dev-server to open the browser after server had been started.
     open: true,
     historyApiFallback: true,
     static: [
       {
-        directory: path.join(appDirectory, 'web/public'),
+        directory: path.join(projectDirectory, 'web/public'),
       },
     ],
     // Hot reload on source changes