+++ nielsen/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Added `useNielsen` hook to create and initialize a Nielsen connector. See README.md for usage.
+
 ## [0.6.0] - 2023-04-14
 
 ### Fixed